+++ .env
@@ -1,0 +1,5 @@
+SSO_START_URL=https://d-9267765094.awsapps.com/start/
+ACCOUNT_ID=763872185484
+ROLE_NAME=AWSReservedSSO_AdministratorAccess_bb7dcaaaba3382bb
+SSO_REGION=us-west-2
+REGION=us-west-2
