+++ lib/get_properties.py
